requirements.txt
fastapi[standard]
uvicorn[standard]
httpx
feedparser
beautifulsoup4
praw
asyncpraw
supabase
python-dotenv
apscheduler
scikit-learn
sentence-transformers
numpy
google-generativeai
discord.py
python-telegram-bot
resend
passlib[bcrypt]
pydantic
pydantic-settings
